Retiring old code base and adding new frontend.
Everything that used to be in this repo has been moved into the `old` folder,
which will be removed in the not-too-distant future. The new readme describes
the scope of the new code, but nothing exists for the new backend yet. This
commit also includes the first version of the new frontend in the `frontend`
directory.