geonames.py now updates database
Currently only adding ADM1 entities. Need to get some
quality measures to determine what others might be best,
and ideally it'd be good to drop terms that never appear (there
seems to be a lot of other languages in the corpus which won't
be used at this point). Won't be a part of this script though.